fix: remove spectralis file
